added new transformers and refactored the current one
