+++ TotemTransformers/totem-general.transformers.cs
@@ -1,0 +1,11 @@
+using System;
+
+namespace TotemTransformers
+{
+    public class GeneralTotemTransformers {
+        public static DateTime ItemDateToDateTime(string dateStr)
+        {
+            return DateTime.Parse(dateStr);
+        }
+    }
+}
