+++ TotemTransformers/totem-item.transformers.cs
@@ -1,9 +1,8 @@
 using UnityEngine;
 
-namespace transformers
+namespace TotemTransformers
 {
-    public class TotemItemTransformer
-    {
+    public class TotemItemTransformer {
         public static Color EntityColorToUnityColor(ColorEntity color)
         {
             return new Color(color.red, color.green, color.blue, color.alpha);
